♻️ 리팩토링: 인터뷰 시뮬레이터 파일 분리
- 인터뷰 클라이언트 로직을 컨트롤러 훅으로 분리\n- View를 패널/슬롯 단위 컴포넌트로 분해\n- 생성 페이지 보안 안내 문구 정리
